Refactor generation copy update scripts
This slightly changes the behaviour of source updating. Previously if
the template source was updated, updating the generated file would
fail. Now the update will fail only if the user has modified the
generated file since it was last generated. This is to detect if user
changes are about to be clobbered as we are modifying source files
outside of the build directory.