template: Implement new templating mechanisms
You will likely need to read the comments on the python modules and
functions to understand all of the changes in this commit (Sorry).

Adds a data model for internal tutorial state that mostly consistes of a
list of tasks with code snippets and completion texts. The point of
creating this data model is to be able to better track each step in a
tutorial for regression and usability purposes, and to embed code
snippets into the actual tutorial documentation rather than duplicating
it, or having documentation that mention any snippets at all.

Adds helper macros and template functions/filters for accessing the
above data model as a tutorial is written.