Change tutorial scripts to use CapDL allocators
This changes the backing data structures for the tutorials that depend on
CapDL to use the Allocators that capDL provides instead of custom ones.
The intention is for this to be more intuitive to use as it is same
patterns of allocation as in Camkes.

We still use custom structures for tracking named CSlots and for
generating region symbols as these features aren't currently provided by
the CapDL structures. If these structures work well, then they may be
incorporated into the CapDL python scripts and used in Camkes in the
future.