capdl_linker: Change file format types
In order to use the capdl_linker in camkes as well as the tutorials, we
serialise the same object type. The tutorials also use the
capdl_linker.py to generate a C file with special symbols defined to
refer to seL4 capabilities and shared memory regions. We change this
input format to take a yaml file to help with readability.