tools: Change spawn -> spawnu in expect.py
In our case, pexpect.spawn outputs bytes. This is a problem with some of
the other tools that we use, especially when we try to write the output
of a spawned process to a file (Python3 doesn't like writing bytes to
files, only strings). Thus we change spawn to spawnu to ensure that
unicode (`str`) is used.