CMake: Refactor tutorial build system structure
Use the new modular build system structure for the tutorials where each
tutorial project is a toplevel project that sets up the rest of the seL4
build system. Now, other than the tooling for regenerating the tutorial
files, the build files for each tutorial can be used as a base for other
applications.  This makes it possible to take a finished tutorial and
keep extending it, checking it into source control and more easily
integrating it with other existing modules.