camkes-vm-crossvm: Cleanup and consolidate tasks
Previously, the tasks were split into individual tasks. This was fine
for the completed state of the tutorials but not for the starting state
of the tutorials. The content for the tasks would not be generated for
the starting state as the tasks were not being advanced. Thus to fix
this, we consolidate all the tasks into one single task but leave each
part as a subtask.