dynamic-2: Fix code block indentation
Some of the code blocks are being indented when we do not want them to.
This is because of a lack of whitespacing between the URL lists and the
start of the code which causes the renderer to think that the code block
belongs to the list. We add whitespace to separate the lists and the
code block so that the renderer will treat the two as non-related
entities.