tutorials: Include a driver for the timer tutes
The CAmkES timer tutorial and the interrupt tutorial
(hello-camkes-timer, interrupts) is susceptible to breakages caused by
ongoing changes to the driver framework and the timer drivers. Thus,
instead of relying on the timer drivers in libplatsupport in the
util_libs repository, we bundle a basic timer driver with the tutorial
sources to make the tutorial more resilient to changes.