Correctly calculate the number of EndpointObjets
The number of EndpointObjects to fill the entire endpoint object should be the size of the untyped object *divided* by the size of the EndpointObjects. The code previous to this PR works, but it does not fill up the object completely, creating a misleading tutorial description.

This bug in the tutorial may be a typo, since for example, `/` and `-` are the same key on a German keyboard with US layout.

I'm just proposing a change to this Markdown file and hope that the change will magically propagate to the code, or do I need to change the actual code somewhere, too?